[REFACT] MainCategoryButton 컴포넌트로 분리 #NADA-49
